ai_api/webapi/requirements.txt
clickhouse_connect
pillow
salesforce-lavis
transformers
fastapi
pydantic
uvicorn
requests
jinja2
python-multipart